Names are being displayed correctly and without repeteition
I added several console.logs to help me understand what is going on in
the for loopand the for each loop.
I am not quite sure why some index positions in the array "people" are
negative numbers.

If any doubt about what is happening appears, look ate the console.logs
in the for loop and for each loop in the main,js